% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunInvBag.R
\name{RunInvBag}
\alias{RunInvBag}
\title{Inverse Bagging algorithm}
\usage{
RunInvBag(train, test, Q = NA, Mboot2 = 10000, R = 1)
}
\arguments{
\item{train}{data frame - Data generated by a certain distribution (so-called background).}

\item{test}{data frame - Data mostly generated by the background with possible signal observations.}

\item{Q}{numeric - Size of sampled sets, which cannot be larger than the test data size}

\item{R}{numeric - Method for scores computation:
\itemize{
\item 1 - Test statistic score
\item 2 - P-value score
\item 3 - OK scores
}}

\item{Mboot}{numeric - Number of sampling iterations.}
}
\value{
A numeric vector of scores for respective observations from the test data set.
}
\description{
The package is an R implementation of the Inverse Bagging algorithm. 
Given two sets: a training one containing observations from a certain distribution 
and a testing one which can include as well anomalous observations generated 
from a different unknown distribution. 
The algorithm returns scores for the testing set observations relating to their anomalous evidence.
}
\examples{
#Generate multivariate data
require(mvtnorm)
set.seed(1)
n <- 2000
B <- 0.05
P <- 4
train <- rmvnorm(n, sigma = diag(P))

#add anomalies with mean shited from sero
test <- rbind(
 rmvnorm(n*(1-B), sigma = diag(P)), 
 rmvnorm(n*B, mean = rep(2,P),sigma = diag(P))
)

#Compute the score
results <- RunInvBag(train, test)

#Plot results
plot(rersults)
}
\references{
Vischia, P. and Tommaso D., "The Inverse Bagging Algorithm: Anomaly Detection by Inverse Bootstrap Aggregating." EPJ Web of Conferences. Vol. 137. EDP Sciences, 2017.
}
